moved fuzzy_match to separate module
